Extract+dedupe `DuplicateHashKeyCheck`
